requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
Django==4.2.16
django-apscheduler==0.7.0
feedparser==6.0.11
python-dateutil==2.9.0.post0
pytz==2024.2
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzlocal==5.2
